FIX problem

+++ includes/raycasting.h
@@ -6,7 +6,7 @@
 /*   By: rpottier <rpottier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/30 09:58:10 by nburat-d          #+#    #+#             */
-/*   Updated: 2022/06/30 16:32:58 by rpottier         ###   ########.fr       */
+/*   Updated: 2022/07/01 13:49:43 by rpottier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,6 +19,6 @@ void	get_shortest_distance(t_ray *ray, t_player *player);
 void	set_facing_values(t_ray *ray);
 void	get_vertical_hit(t_ray *ray, t_player *player, t_map *map);
 void	get_horizontal_hit(t_ray *ray, t_player *player, t_map *map);
-t_ray	*cast_ray(float angle, t_player *player, t_map *map);
+t_ray	*cast_ray(double angle, t_player *player, t_map *map);
 
 #endif
