+++ srcs/parsing/set_parameters.c
@@ -6,7 +6,7 @@
 /*   By: rpottier <rpottier@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/21 10:25:19 by rpottier          #+#    #+#             */
-/*   Updated: 2022/06/29 17:01:00 by rpottier         ###   ########.fr       */
+/*   Updated: 2022/07/01 13:49:43 by rpottier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,7 +38,7 @@ t_data	*set_parameters(t_file	*file)
 	return (data);
 }
 
-float	get_start_angle(char c)
+double	get_start_angle(char c)
 {
 	if (c == 'E')
 		return (0);
